Cleaned Up Project and Added Update Logo Page
There were a bunch of junk files that came with the new asp.net project
that we weren't going to need, so I removed them.

When they are there, I'll go ahead and add in the code for showing the
correct logo onto your pages.

I added ADMIN_LOGONAME as a field into the ADMINISTRATOR table. It
tracks the correct filename and checks for its existence every time a
page loads.